fixing error in Sling Models test code - service ranking must be an integer to be effective
git-svn-id: https://svn.apache.org/repos/asf/sling/trunk/bundles/extensions/models/impl@1619818 13f79535-47bb-0310-9956-ffa450edef68

+++ src/test/java/org/apache/sling/models/impl/ServicePropertiesMap.java
@@ -23,7 +23,7 @@ import org.osgi.framework.Constants;
 @SuppressWarnings("serial")
 public class ServicePropertiesMap extends HashMap<String, Object> {
     
-    public ServicePropertiesMap(long serviceId, long serviceRanking) {
+    public ServicePropertiesMap(long serviceId, int serviceRanking) {
         super();
         put(Constants.SERVICE_ID, serviceId);
         put(Constants.SERVICE_RANKING, serviceRanking);
